+++ src/css/components/pane.less
@@ -47,6 +47,10 @@
             margin-left: @horizontal-spacing-unit;
           }
         }
+
+        .list-inline > li {
+          padding-right: 0;
+        }
       }
     }
