+++ src/js/components/SidebarComponent.jsx
@@ -1,4 +1,3 @@
-import {Link} from "react-router";
 import React from "react/addons";
 
 import FilterTypes from "../constants/FilterTypes";
@@ -29,25 +28,8 @@ var SidebarComponent = React.createClass({
   },
 
   render: function () {
-    var path = this.context.router.getCurrentPathname();
-    var props = this.props;
-
-    var newAppModalQuery = {
-      modal: "new-app"
-    };
-
-    if (props.groupId != null && props.groupId !== "/") {
-      newAppModalQuery.groupId = props.groupId;
-    }
-
     return (
       <nav className="sidebar">
-        <Link to={path}
-            query={newAppModalQuery}
-            className="btn btn-success create-app"
-            activeClassName="create-app-active">
-          Create
-        </Link>
         <div className="flex-row">
           <h3 className="small-caps">Status</h3>
           {this.getClearLinkForFilter(FilterTypes.STATUS)}
